last working state. All classes can be inputted

+++ app/src/main/java/com/example/tenchrio/flashcardz/database/ChapterDAO.java
@@ -19,7 +19,7 @@ public class ChapterDAO {
         fcDbHelper = FlashCardDBHelper.getInstance(context);
     }
 
-    public Chapter addChapter(Chapter newChapter, Integer courseID) {
+    public Chapter addChapter(Chapter newChapter, long courseID) {
         SQLiteDatabase db = fcDbHelper.getWritableDatabase();
 
         ContentValues values = new ContentValues();
@@ -32,17 +32,14 @@ public class ChapterDAO {
         return newChapter;
     }
 
-    public List<Chapter> getAllChaptersForCourse(int CourseID)
+    public List<Chapter> getAllChaptersForCourse(long CourseID)
     {
-        // Gets the data repository in read mode
         SQLiteDatabase db = fcDbHelper.getReadableDatabase();
 
-        // Create SQL query
         String selectQuery = "SELECT * FROM " +
                 DatabaseContract.ChapterEntry.TABLE_NAME +
                 " WHERE " + DatabaseContract.ChapterEntry.COLUMN_NAME_COURSEID + " = " + CourseID;
 
-        // Execute query, this will return a cursor that can be used to iterate results
         Cursor c = db.rawQuery(selectQuery, null);
         ArrayList<Chapter> chapters= new ArrayList<>();
         if (c != null) {
